Update priority validation in task webhook schemas to enforce integer values
- Changed priority validation from `z.number().min(0).max(3)` to `z.number().int().min(0).max(3)` in both create and update task webhook body schemas.
- This ensures that only integer values are accepted for task priority, maintaining data integrity and consistency.

+++ api/index.ts
@@ -94,7 +94,7 @@ const createTaskWebhookBodySchema = z.object({
   title: z.string().min(1).trim(),
   notes: z.string().trim().optional(),
   due: z.string().optional(),
-  priority: z.number().min(0).max(3).optional(), // 0: None, 1: Low, 2: Medium, 3: High
+  priority: z.number().int().min(0).max(3).optional(), // 0: None, 1: Low, 2: Medium, 3: High
   url: z.string().optional(),
   tags: z.string().optional(),
   syncId: z.string().optional(),
@@ -107,7 +107,7 @@ const updateTaskWebhookBodySchema = z
     title: z.string().min(1).trim().optional(),
     notes: z.string().trim().optional(),
     due: z.string().optional(),
-    priority: z.number().min(0).max(3).optional(), // 0: None, 1: Low, 2: Medium, 3: High
+    priority: z.number().int().min(0).max(3).optional(), // 0: None, 1: Low, 2: Medium, 3: High
     isCompleted: z.boolean().optional(),
     url: z.string().optional(),
     tags: z.string().optional(),
